chore(*) generate system resource wrappers
