+++ api/system/v1alpha1/config.proto
@@ -4,9 +4,18 @@ package kuma.system.v1alpha1;
 
 option go_package = "github.com/kumahq/kuma/api/system/v1alpha1";
 
+import "mesh/options.proto";
+
 // Config is a entity that represents dynamic configuration that is stored in
 // underlying storage. For now it's used only for internal mechanisms.
 message Config {
+
+  option (kuma.mesh.resource).name = "ConfigResource";
+  option (kuma.mesh.resource).type = "Config";
+  option (kuma.mesh.resource).package = "system";
+  option (kuma.mesh.resource).global = true;
+  option (kuma.mesh.resource).skip_validation = true;
+
   // configuration that is stored (ex. in JSON)
   string config = 1;
 }
