+++ api/system/v1alpha1/secret.proto
@@ -4,10 +4,17 @@ package kuma.system.v1alpha1;
 
 option go_package = "github.com/kumahq/kuma/api/system/v1alpha1";
 
+import "mesh/options.proto";
 import "google/protobuf/wrappers.proto";
 
-// Secret defines encrypted value in Kuma
+// Secret defines an encrypted value in Kuma.
 message Secret {
+
+  option (kuma.mesh.resource).name = "SecretResource";
+  option (kuma.mesh.resource).type = "Secret";
+  option (kuma.mesh.resource).package = "system";
+  option (kuma.mesh.resource).skip_validation = true;
+
   // Value of the secret
   google.protobuf.BytesValue data = 1;
 }
