+++ api/system/v1alpha1/zone.proto
@@ -4,11 +4,18 @@ package kuma.system.v1alpha1;
 
 option go_package = "github.com/kumahq/kuma/api/system/v1alpha1";
 
+import "mesh/options.proto";
 import "google/protobuf/wrappers.proto";
 
 // Zone defines the Zone configuration used at the Global Control Plane
 // within a distributed deployment
 message Zone {
+
+  option (kuma.mesh.resource).name = "ZoneResource";
+  option (kuma.mesh.resource).type = "Zone";
+  option (kuma.mesh.resource).package = "system";
+  option (kuma.mesh.resource).global = true;
+
   // enable allows to turn the zone on/off and exclude the whole zone from
   // balancing traffic on it
   google.protobuf.BoolValue enabled = 1;
