+++ tools/resource-gen.go
@@ -231,12 +231,12 @@ package %s
 import (
 	"fmt"
 
-	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
+	mesh_proto "github.com/kumahq/kuma/api/%s/v1alpha1"
 	"github.com/kumahq/kuma/pkg/core/resources/model"
 	"github.com/kumahq/kuma/pkg/core/resources/registry"
 )
 
-`, Package)
+`, Package, Package)
 
 	var types []protoreflect.MessageType
